+++ src/main/java/sic/repository/NotaDebitoRepository.java
@@ -1,5 +1,6 @@
 package sic.repository;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -15,7 +16,7 @@ public interface NotaDebitoRepository extends NotaRepository<NotaDebito> {
     NotaDebito getById(@Param("idNotaDebito") long idNotaDebito);
     
     @Query("SELECT SUM(nd.total) FROM NotaDebito nd WHERE nd.empresa = :empresa AND nd.cliente = :cliente AND nd.eliminada = false AND nd.fecha <= :hasta")
-    Double totalNotasDebito(@Param("hasta") Date hasta, @Param("cliente") Cliente cliente, @Param("empresa") Empresa empresa);
+    BigDecimal totalNotasDebito(@Param("hasta") Date hasta, @Param("cliente") Cliente cliente, @Param("empresa") Empresa empresa);
     
     NotaDebito findTopByEmpresaAndTipoComprobanteOrderByNroNotaDesc(Empresa empresa, TipoDeComprobante tipoComprobante);
