+++ src/main/java/sic/service/IProductoService.java
@@ -1,5 +1,6 @@
 package sic.service;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import sic.modelo.TipoDeOperacion;
 import java.util.List;
@@ -17,27 +18,27 @@ public interface IProductoService {
 
     void actualizar(Producto producto);
 
-    void actualizarStock(HashMap<Long, Double> idsYCantidades, TipoDeOperacion operacion, Movimiento movimiento);
+    void actualizarStock(HashMap<Long, BigDecimal> idsYCantidades, TipoDeOperacion operacion, Movimiento movimiento);
 
     Page<Producto> buscarProductos(BusquedaProductoCriteria criteria);
 
-    double calcularGanancia_Neto(double precioCosto, double ganancia_porcentaje);
+    BigDecimal calcularGanancia_Neto(BigDecimal precioCosto, BigDecimal ganancia_porcentaje);
 
-    Map<Long, Double> getProductosSinStockDisponible(long[] idProducto, double[] cantidad);
+    Map<Long, BigDecimal> getProductosSinStockDisponible(long[] idProducto, BigDecimal[] cantidad);
 
-    Map<Long, Double> getProductosNoCumplenCantidadVentaMinima(long[] idProducto, double[] cantidad);
+    Map<Long, BigDecimal> getProductosNoCumplenCantidadVentaMinima(long[] idProducto, BigDecimal[] cantidad);
     
-    double calcularGanancia_Porcentaje(Double precioDeListaNuevo, 
-            Double precioDeListaAnterior, double pvp, Double ivaPorcentaje, 
-            Double impInternoPorcentaje, double precioCosto, boolean descendente);
+    BigDecimal calcularGanancia_Porcentaje(BigDecimal precioDeListaNuevo, 
+            BigDecimal precioDeListaAnterior, BigDecimal pvp, BigDecimal ivaPorcentaje, 
+            BigDecimal impInternoPorcentaje, BigDecimal precioCosto, boolean descendente);
 
-    double calcularIVA_Neto(double precioCosto, double iva_porcentaje);
+    BigDecimal calcularIVA_Neto(BigDecimal precioCosto, BigDecimal iva_porcentaje);
 
-    double calcularImpInterno_Neto(double precioCosto, double impInterno_porcentaje);
+    BigDecimal calcularImpInterno_Neto(BigDecimal precioCosto, BigDecimal impInterno_porcentaje);
 
-    double calcularPVP(double precioCosto, double ganancia_porcentaje);
+    BigDecimal calcularPVP(BigDecimal precioCosto, BigDecimal ganancia_porcentaje);
 
-    double calcularPrecioLista(double PVP, double iva_porcentaje, double impInterno_porcentaje);    
+    BigDecimal calcularPrecioLista(BigDecimal PVP, BigDecimal iva_porcentaje, BigDecimal impInterno_porcentaje);    
 
     void eliminarMultiplesProductos(long[] idProducto);
 
@@ -47,7 +48,7 @@ public interface IProductoService {
 
     Producto getProductoPorId(long id_Producto);
     
-    double calcularValorStock(BusquedaProductoCriteria criteria);
+    BigDecimal calcularValorStock(BusquedaProductoCriteria criteria);
   
     byte[] getReporteListaDePreciosPorEmpresa(List<Producto> productos, Empresa empresa);
 
@@ -55,15 +56,15 @@ public interface IProductoService {
 
     List<Producto> modificarMultiplesProductos(long[] idProducto,
                                                boolean checkPrecios,            
-                                               Double gananciaNeto,
-                                               Double gananciaPorcentaje,
-                                               Double impuestoInternoNeto,
-                                               Double impuestoInternoPorcentaje,
-                                               Double IVANeto,
-                                               Double IVAPorcentaje,
-                                               Double precioCosto,
-                                               Double precioLista,
-                                               Double precioVentaPublico,                                                                     
+                                               BigDecimal gananciaNeto,
+                                               BigDecimal gananciaPorcentaje,
+                                               BigDecimal impuestoInternoNeto,
+                                               BigDecimal impuestoInternoPorcentaje,
+                                               BigDecimal IVANeto,
+                                               BigDecimal IVAPorcentaje,
+                                               BigDecimal precioCosto,
+                                               BigDecimal precioLista,
+                                               BigDecimal precioVentaPublico,                                                                     
                                                boolean checkMedida, Medida medida,
                                                boolean checkRubro, Rubro rubro,
                                                boolean checkProveedor, Proveedor proveedor);
